+++ CHANGELOG.md
@@ -11,7 +11,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added Opensearch serverless module
+- updated replication module to avoid docker pull rate limits and resource creation race conditions
+
 ### **Changed**
+
+### **Removed**
+
+=======
+
+=======
+
+## v1.3.0 (2024-01-16)
+
+### **Added**
+
+- added eks node iam role that all eks nodes will assume on start
+- added support for S3 CSI driver
+
+### **Changed**
+
+- added logic to require IMDSv2 in eks nodes
+- regrouped ecr module from `containers` to `storage` group
+- fixed the execution of workflow logic to remove an additional `/`
 
 ### **Removed**